chore: rename cache util to cache helper
