+++ backend/tests/unit/utils/test_matches_cache_helper.py
@@ -15,7 +15,7 @@ from app.constants.cache_ttl import (
     MATCHES_SOON_SCHEDULED_TTL,
 )
 from app.models.match import StatusType
-from app.utils.cache import get_matches_ttl
+from backend.app.utils.cache_helper import get_matches_ttl
 
 CURRENT_TIME = datetime(2026, 6, 15, 12, 0, tzinfo=UTC)
